+++ js/lib.js
@@ -1,6 +1,6 @@
 /* *********************************************************************
   Threepenny - JavaScript library
-  
+
   Additional JavaScript functions that are available
   for Threepenny.
 ********************************************************************* */
@@ -34,7 +34,7 @@ Haskell.bind = function (el, eventType, fun) {
             ? [e.originalEvent.dataTransfer.getData("dragData")]
             : [] );
     });
-  } else if(eventType.match('mousemove|mousedown|mouseup')) {
+  } else if(eventType.match('contextmenu|mousemove|mousedown|mouseup')) {
     $(el).bind(eventType, function(e) {
       var offset = $(this).offset();
       var x      = e.pageX - offset.left;
@@ -58,7 +58,7 @@ Haskell.bind = function (el, eventType, fun) {
 /////////////////////////////////////////////////////////////////////
 // Canvas API additions.
 // See http://stackoverflow.com/a/9722502/403805 .
-CanvasRenderingContext2D.prototype.clear = 
+CanvasRenderingContext2D.prototype.clear =
   CanvasRenderingContext2D.prototype.clear || function (preserveTransform) {
     if (preserveTransform) {
       this.save();
@@ -67,7 +67,7 @@ CanvasRenderingContext2D.prototype.clear =
     this.clearRect(0, 0, this.canvas.width, this.canvas.height);
     if (preserveTransform) {
       this.restore();
-    }           
+    }
 };