+++ samples/GetElementsBy.hs
@@ -14,13 +14,14 @@ setup w = do
 
     button1 <- UI.button # set UI.text "by tag"
     button2 <- UI.button # set UI.text "by class"
-    button3 <- UI.button # set UI.text "by id"    # set UI.id_ "me"
+    button3 <- UI.button # set UI.text "by id 'me'" # set UI.id_ "me"
+    button4 <- UI.button # set UI.text "by id 'ma'"
 
     let mkString s = UI.string s # set UI.class_ "string"
 
     getBody w #+ [column $
-        row [element button1, element button2, element button3]
-        : map mkString (words "We are so different")]
+        row [element button1, element button2, element button3, element button4]
+        : map mkString (words "Many different elements")]
 
     on UI.click button1 $ const $ do
         xs <- getElementsByTagName w "span"
@@ -38,3 +39,9 @@ setup w = do
         let s = if b then "yay" else "wow"
         element x # set UI.text s
         liftIO $ writeIORef ref (not b)
+
+    on UI.click button4 $ const $ void $ do
+        m <- getElementById w "ma"
+        case m of
+            Nothing -> element button4 # set UI.text "not found"
+            Just el -> element el      # set UI.text "found"